+++ x/poolmanager/client/cli/query.go
@@ -32,6 +32,7 @@ func GetQueryCmd() *cobra.Command {
 	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetCmdPool)
 	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetCmdTotalVolumeForPool)
 	osmocli.AddQueryCmd(cmd, queryprotov2.NewQueryClient, GetCmdSpotPriceV2)
+	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetCmdEstimateTradeBasedOnPriceImpact)
 	cmd.AddCommand(
 		osmocli.GetParams[*queryproto.ParamsRequest](
 			types.ModuleName, queryproto.NewQueryClient),
@@ -104,16 +105,6 @@ func GetCmdSpotPrice() (*osmocli.QueryDescriptor, *queryproto.SpotPriceRequest) 
 {{.CommandPrefix}} spot-price 1 uosmo ibc/27394FB092D2ECCD56123C74F36E4C1F926001CEADA9CA97EA622B25F41E5EB2
 `,
 	}, &queryproto.SpotPriceRequest{}
-}
-
-func GetCmdSpotPriceV2() (*osmocli.QueryDescriptor, *queryprotov2.SpotPriceRequest) {
-	return &osmocli.QueryDescriptor{
-		Use:   "spot-price-v2",
-		Short: "Query spot-price with 36 decimals",
-		Long: `Query spot-price with 36 decimals
-{{.CommandPrefix}} spot-price-v2 1 uosmo ibc/27394FB092D2ECCD56123C74F36E4C1F926001CEADA9CA97EA622B25F41E5EB2
-`,
-	}, &queryprotov2.SpotPriceRequest{}
 }
 
 func EstimateSwapExactAmountInParseArgs(args []string, fs *flag.FlagSet) (proto.Message, error) {
@@ -190,4 +181,16 @@ func GetCmdTotalVolumeForPool() (*osmocli.QueryDescriptor, *queryproto.TotalVolu
 		Long: `{{.Short}}
 		{{.CommandPrefix}} total-volume-for-pool 1`,
 	}, &queryproto.TotalVolumeForPoolRequest{}
+}
+
+func GetCmdEstimateTradeBasedOnPriceImpact() (
+	*osmocli.QueryDescriptor, *queryproto.EstimateTradeBasedOnPriceImpactRequest,
+) {
+	return &osmocli.QueryDescriptor{
+		Use:   "estimate-trade-based-on-price-impact  <fromCoin> <toCoinDenom> <poolId> <maxPriceImpact> <externalPrice>",
+		Short: "Query estimate-trade-based-on-price-impact",
+		Long: `{{.Short}} 
+		{{.CommandPrefix}} estimate-trade-based-on-price-impact 100uosmo stosmo  833 0.001 1.00`,
+		QueryFnName: "EstimateTradeBasedOnPriceImpact",
+	}, &queryproto.EstimateTradeBasedOnPriceImpactRequest{}
 }